stateful-agents/requirements.txt
langchain>=0.1.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
# Jupyter and IPython dependencies
jupyter>=1.0.0
ipykernel>=6.0.0
ipython>=8.0.0
notebook>=7.0.0
jupyter-console>=6.6.0
qtconsole>=5.4.0 